backend/requirements.txt
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
Django==6.0.1
django-ckeditor-5==0.2.19
django-cors-headers==4.9.0
django-js-asset==3.1.2
djangorestframework==3.16.1
gunicorn==23.0.0
idna==3.11
packaging==25.0
pillow==12.1.0
python-decouple==3.8
python-dotenv==1.2.1
requests==2.32.5
resend==2.21.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
